revisi migration karyawan

+++ database/migrations/2023_05_01_083159_create_karyawans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('m_karyawan', function (Blueprint $table) {
-            $table->integer('kry_id')->unsigned();
+            $table->unsignedBigInteger('kry_id')->unsigned();
             $table->string('kry_nik')->unique();
             $table->string('kry_nama');
             $table->string('kry_bagian');
